Refactor form imports to prevent circular dependencies
