+++ accounting_app/accounting/forms/__init__.py
@@ -1,16 +1,5 @@
-from .partners import (
-    PartnerForm,
-    PartnersGroupForm,
-    PartnersGroupMemberForm,
-    PartnersGroupMemberFormSet,
-)
-from .safes import SafeForm
-from .customers import CustomerForm
-from .contracts import ContractForm
-from .vouchers import ReceiptVoucherForm, PaymentVoucherForm
-from .settlements import SettlementForm
-from .inventory import ItemForm, StockMoveForm
-from .projects import ProjectForm
-from .installments import InstallmentForm, InstallmentPaymentForm
-from .suppliers import SupplierForm
-from .units import UnitForm
+# Forms are imported individually to avoid circular imports
+# Import forms as needed in views using:
+# from accounting.forms.partners import PartnerForm
+# from accounting.forms.customers import CustomerForm
+# etc.
